console: Refactor application add view
Move from class components to functional components
